// src/app/api-client/model/post.ts
/**
 * Blog API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Comment } from './comment';
import { Usuario } from './usuario';


export interface Post { 
    id?: number;
    titulo?: string;
    post?: string;
    data?: string;
    tempo?: string;
    comentarios?: Array<Comment>;
    usuario?: Usuario;
}